Updated cohort functional group defintion file
Contains Adult/juvenile ratio mode which can be set to fixed or scaled:
i.e., if juvenile size will scale with adult body mass or if it is
always the same. And, Adult/juvenile ratio which in the case of scaled
mode will define the size of the juvenile based on the size of the adult
or in the case of fixed mode will set the size of the juvenile
(independetly from adult size).